fix(security): gate repo rule files, centralize state dir, harden persistence
A cloned repo could still steer tokenix without `tokenix trust`: repo
secret-rules and egress-rules loaded unconditionally, so an added `.+`
rule would blank every PostToolUse result. They now load only once the
repo is trusted and stay add-only. Memory notes are also checked against
the bundled secret rules, so bare AKIA/ghp_ tokens are refused.

Persisted views (recordings, unfiltered log, snapshots) are redacted, the
daemon status endpoint needs the capability token, re-read suppression is
scoped to the agent session, and filter generate asks before re-running.

store::global_dir() is now the single source of ~/.tokenix and honours an
absolute TOKENIX_HOME, so tests and verify.sh never write the real home.
find_project_root prefers an indexed directory over marker files: a
package.json in a home directory captured every marker-less repo below it.

search_similar scans vectors only and loads text for the top k
(122-130 ms -> 47-48 ms/query on a 30k x 768d bench). fastembed 4.9 -> 7
and ort rc.13 (models behind a Mutex, renamed EPs); nomic vectors are
bit-identical, so existing indexes stay valid. rustls bumped for
RUSTSEC-2026-0285. Dependency and code changes share embed.rs, so they
ship together.

+++ src/graph.rs
@@ -888,7 +888,7 @@ fn extract_references_tree_sitter(content: &str, path: &str) -> Option<Vec<Strin
             }
         }
 
-        for i in 0..node.child_count() as u32 {
+        for i in 0..node.child_count() {
             if let Some(child) = node.child(i) {
                 traverse(child, content, refs);
             }
